pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=1.0.5",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "langchainhub==0.1.21",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
]
